add caching mechanism
this PR restructures experiment I/O as follows.

As before, we save each run outcome in a json file.
There is one significant change: we now ensure that each json file only includes a single attack run (top level is now a dict rather than a list).

In addition, there is one tinydb database which stores a map of config -> (path, judge status).

Before a run begins, we cross-check against the db and only run the experiments which we have not run so far.

TODO: have judges add their status